# contracts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "简历随行私有 API",
    "version": "0.4.0"
  },
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "summary": "Capabilities",
        "operationId": "capabilities_v1_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            }
          },
          "400": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "408": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "499": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerToken": []
          },
          {
            "WebSession": []
          }
        ]
      }
    },
    "/v1/readiness": {
      "get": {
        "summary": "Readiness",
        "operationId": "readiness_v1_readiness_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyResponse"
                }
              }
            }
          },
          "400": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "408": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "499": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerToken": []
          },
          {
            "WebSession": []
          }
        ]
      }
    },
    "/v1/source-catalog": {
      "get": {
        "summary": "Source Catalog",
        "operationId": "source_catalog_v1_source_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogResponse"
                }
              }
            }
          },
          "400": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "408": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "499": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerToken": []
          },
          {
            "WebSession": []
          }
        ]
      }
    },
    "/v1/fields/resolve": {
      "post": {
        "summary": "Resolve",
        "operationId": "resolve_v1_fields_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            }
          },
          "400": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "408": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "499": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "统一错误；不回显请求内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerToken": []
          },
          {
            "WebSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CapabilitiesResponse": {
        "properties": {
          "protocol_version": {
            "type": "string",
            "const": "1.0",
            "title": "Protocol Version"
          },
          "mode": {
            "type": "string",
            "enum": [
              "mock",
              "private_model"
            ],
            "title": "Mode"
          },
          "features": {
            "$ref": "#/components/schemas/Features"
          },
          "limits": {
            "$ref": "#/components/schemas/Limits"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "protocol_version",
          "mode",
          "features",
          "limits"
        ],
        "title": "CapabilitiesResponse"
      },
      "CatalogResponse": {
        "properties": {
          "protocol_version": {
            "type": "string",
            "const": "1.0",
            "title": "Protocol Version"
          },
          "source_catalog": {
            "items": {
              "$ref": "#/components/schemas/SourceItem"
            },
            "type": "array",
            "title": "Source Catalog"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "protocol_version",
          "source_catalog"
        ],
        "title": "CatalogResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "Features": {
        "properties": {
          "resolve_fields": {
            "type": "boolean",
            "title": "Resolve Fields"
          },
          "draft_answer": {
            "type": "boolean",
            "title": "Draft Answer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "resolve_fields",
          "draft_answer"
        ],
        "title": "Features"
      },
...
      "Limits": {
        "properties": {
          "max_body_bytes": {
            "type": "integer",
            "title": "Max Body Bytes"
          },
          "max_fields": {
            "type": "integer",
            "title": "Max Fields"
          },
          "max_facts": {
            "type": "integer",
            "title": "Max Facts"
          },
          "max_answer_characters": {
            "type": "integer",
            "title": "Max Answer Characters"
          },
          "max_input_characters": {
            "type": "integer",
            "title": "Max Input Characters"
          },
          "request_deadline_seconds": {
            "type": "number",
            "title": "Request Deadline Seconds"
          },
          "max_concurrent": {
            "type": "integer",
            "title": "Max Concurrent"
          },
          "max_concurrent_per_token": {
            "type": "integer",
            "title": "Max Concurrent Per Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "max_body_bytes",
          "max_fields",
          "max_facts",
          "max_answer_characters",
          "max_input_characters",
          "request_deadline_seconds",
          "max_concurrent",
          "max_concurrent_per_token"
        ],
        "title": "Limits"
      },
...
      "ReadyResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ready",
            "title": "Status"
          },
          "mode": {
            "type": "string",
            "enum": [
              "mock",
              "private_model"
            ],
            "title": "Mode"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "mode",
          "model"
        ],
        "title": "ReadyResponse"
      },
...
      "ResolveRequest": {
        "properties": {
          "protocol_version": {
            "type": "string",
            "const": "1.0",
            "title": "Protocol Version"
          },
          "request_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,100}$",
            "title": "Request Id"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/WebField"
            },
            "type": "array",
            "maxItems": 40,
            "minItems": 1,
            "title": "Fields"
          },
          "source_catalog": {
            "items": {
              "$ref": "#/components/schemas/SourceItem"
            },
            "type": "array",
            "maxItems": 120,
            "minItems": 1,
            "title": "Source Catalog"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "protocol_version",
          "request_id",
          "fields",
          "source_catalog"
        ],
        "title": "ResolveRequest"
      },
      "ResolveResponse": {
        "properties": {
          "protocol_version": {
            "type": "string",
            "const": "1.0",
            "title": "Protocol Version"
          },
          "request_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,100}$",
            "title": "Request Id"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/FieldResult"
            },
            "type": "array",
            "maxItems": 40,
            "title": "Results"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "protocol_version",
          "request_id",
          "results"
        ],
        "title": "ResolveResponse"
      },
...
    },
    "securitySchemes": {
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "个人 API Key；仅可读取已确认简历或调用获授权的模型接口"
      },
      "WebSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "resume_session",
        "description": "网页登录会话；写操作同时携带 X-CSRF-Token"
      }
    }
  }
}